Refactor out CommonApiClient
