+++ src/DigitalPreservation/DigitalPreservation.sln
@@ -30,6 +30,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DigitalPreservation.Common.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DigitalPreservation.Utils", "DigitalPreservation.Utils\DigitalPreservation.Utils.csproj", "{3030ABE2-C922-4E07-9D4E-5A86383E941D}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DigitalPreservation.CommonApiClient", "DigitalPreservation.CommonApiClient\DigitalPreservation.CommonApiClient.csproj", "{D4591C54-73DD-4277-96D3-16468A99BBC0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -92,6 +94,10 @@ Global
 		{3030ABE2-C922-4E07-9D4E-5A86383E941D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3030ABE2-C922-4E07-9D4E-5A86383E941D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3030ABE2-C922-4E07-9D4E-5A86383E941D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D4591C54-73DD-4277-96D3-16468A99BBC0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D4591C54-73DD-4277-96D3-16468A99BBC0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D4591C54-73DD-4277-96D3-16468A99BBC0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D4591C54-73DD-4277-96D3-16468A99BBC0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{E045329F-268B-4E69-8C2A-82B2CBEB9014} = {2FF7269F-F433-41F9-84BF-EAA8CFCFFCF5}
