+++ src/DigitalPreservation/Preservation.Client/PreservationApiClient.cs
@@ -1,106 +1,21 @@
-﻿using System.Net;
-using System.Net.Http.Json;
-using DigitalPreservation.Common.Model;
-using DigitalPreservation.Common.Model.Results;
-using DigitalPreservation.Utils;
+﻿using System.Net.Http.Json;
+using DigitalPreservation.CommonApiClient;
 using Microsoft.Extensions.Logging;
 using Storage.Repository.Common;
 
 namespace Preservation.Client;
 
-internal class PreservationApiClient(HttpClient httpClient, ILogger<PreservationApiClient> logger) : IPreservationApiClient
+internal class PreservationApiClient(
+    HttpClient httpClient,
+    ILogger<PreservationApiClient> logger) : CommonApiBase(httpClient, logger), IPreservationApiClient
 {
-    public async Task<Result<PreservedResource?>> GetResource(string path)
-    {
-        // THIS CODE IS IDENTICAL TO StorageApiClient !!!
-        // path MUST be the full /repository... path, which we just pass through as-is
-        try
-        {
-            var uri = new Uri(path, UriKind.Relative);
-            var req = new HttpRequestMessage(HttpMethod.Get, uri);
-            var response = await httpClient.SendAsync(req);
-            var stream = await response.Content.ReadAsStreamAsync();
-            if (response.IsSuccessStatusCode)
-            {
-                var parsed = Deserializer.Parse(stream);
-                if (parsed != null)
-                {
-                    return Result.Ok<PreservedResource?>(parsed);
-                }
-                return Result.Fail<PreservedResource?>(ErrorCodes.UnknownError, "Resource could not be parsed.");
-            }
-
-            switch (response.StatusCode)
-            {
-                case HttpStatusCode.NotFound:
-                    return Result.Fail<PreservedResource?>(ErrorCodes.NotFound, "No resource at " + uri);
-                case HttpStatusCode.Unauthorized:
-                    return Result.Fail<PreservedResource?>(ErrorCodes.Unauthorized, "Unauthorized for " + uri);
-                default:
-                    return Result.Fail<PreservedResource?>(ErrorCodes.UnknownError, "Status " + response.StatusCode);
-            }
-        }
-        catch (Exception e)
-        {
-            logger.LogError(e, e.Message);
-            return Result.Fail<PreservedResource?>(ErrorCodes.UnknownError, e.Message);
-        }
-    }
-
-    public async Task<Result<Container?>> CreateContainer(string path, string? name = null)
-    {
-        // THIS CODE IS IDENTICAL TO StorageApiClient !!!
-        try
-        {
-            var uri = new Uri(path, UriKind.Relative);
-            HttpResponseMessage response;
-            if (name.HasText())
-            {
-                var container = new Container { Name = name };
-                response = await httpClient.PutAsJsonAsync(uri, container);
-            }
-            else
-            {
-                response = await httpClient.PutAsync(uri, null);
-            }
-            var stream = await response.Content.ReadAsStreamAsync();
-            
-            if (response.IsSuccessStatusCode)
-            {
-                var parsed = Deserializer.Parse(stream);
-                if (parsed is Container createdContainer)
-                {
-                    return Result.Ok<Container?>(createdContainer);
-                }
-                return Result.Fail<Container?>(ErrorCodes.UnknownError, "Resource could not be parsed.");
-            }
-
-            switch (response.StatusCode)
-            {
-                case HttpStatusCode.Conflict:
-                    return Result.Fail<Container?>(ErrorCodes.Conflict, "Conflicting resource at " + uri);
-                case HttpStatusCode.Unauthorized:
-                    return Result.Fail<Container?>(ErrorCodes.Unauthorized, "Unauthorized for " + uri);
-                case HttpStatusCode.BadRequest:
-                    return Result.Fail<Container?>(ErrorCodes.BadRequest, "Bad Request");
-                default:
-                    return Result.Fail<Container?>(ErrorCodes.UnknownError, "Status " + response.StatusCode);
-            }
-            
-        }
-        catch (Exception e)
-        {
-            logger.LogError(e, e.Message);
-            return Result.Fail<Container?>(ErrorCodes.UnknownError, e.Message);
-        }
-    }
-
+    private readonly HttpClient preservationHttpClient = httpClient;
 
     public async Task<ConnectivityCheckResult?> IsAlive(CancellationToken cancellationToken = default)
     {
         try
         {
-            var res = await httpClient.GetFromJsonAsync<ConnectivityCheckResult>("/storage", cancellationToken);
+            var res = await preservationHttpClient.GetFromJsonAsync<ConnectivityCheckResult>("/storage", cancellationToken);
             return res;
         }
         catch (Exception ex)
@@ -119,7 +34,7 @@ internal class PreservationApiClient(HttpClient httpClient, ILogger<Preservation
     {
         try
         {
-            var res = await httpClient.GetFromJsonAsync<ConnectivityCheckResult>("/storage/check-s3", cancellationToken);
+            var res = await preservationHttpClient.GetFromJsonAsync<ConnectivityCheckResult>("/storage/check-s3", cancellationToken);
             return res;
         }
         catch (Exception ex)
@@ -138,7 +53,7 @@ internal class PreservationApiClient(HttpClient httpClient, ILogger<Preservation
     {
         try
         {
-            var res = await httpClient.GetFromJsonAsync<ConnectivityCheckResult>("/storage/check-storage-s3", cancellationToken);
+            var res = await preservationHttpClient.GetFromJsonAsync<ConnectivityCheckResult>("/storage/check-storage-s3", cancellationToken);
             return res;
         }
         catch (Exception ex)
