+++ src/DigitalPreservation/Storage.Client/Storage.Client.csproj
@@ -13,6 +13,7 @@
 
     <ItemGroup>
       <ProjectReference Include="..\DigitalPreservation.Common.Model\DigitalPreservation.Common.Model.csproj" />
+      <ProjectReference Include="..\DigitalPreservation.CommonApiClient\DigitalPreservation.CommonApiClient.csproj" />
       <ProjectReference Include="..\DigitalPreservation.Core\DigitalPreservation.Core.csproj" />
       <ProjectReference Include="..\Storage.Repository.Common\Storage.Repository.Common.csproj" />
     </ItemGroup>
